- corrected type of DebugMemberLayout.BitSize, it is now ulong, which matches the DataLayout.BitSize - fixed arg error checking on DebugUnionType and enforced LLVM requirement that the union has a DIFile

+++ src/Llvm.NET/DebugInfo/DebugMemberInfo.cs
@@ -15,7 +15,7 @@ namespace Llvm.NET.DebugInfo
         /// <param name="bitSize">Size of the member in bits</param>
         /// <param name="bitAlignment">Alignment of the member in bits</param>
         /// <param name="bitOffset">Offset of the member in bits</param>
-        public DebugMemberLayout( uint bitSize, uint bitAlignment, ulong bitOffset )
+        public DebugMemberLayout( ulong bitSize, uint bitAlignment, ulong bitOffset )
         {
             BitSize = bitSize;
             BitAlignment = bitAlignment;
@@ -23,7 +23,7 @@ namespace Llvm.NET.DebugInfo
         }
 
         /// <summary>Bit size for the field</summary>
-        public uint BitSize { get; }
+        public ulong BitSize { get; }
 
         /// <summary>Bit alignment for the field</summary>
         public uint BitAlignment { get; }
